[modes] split mode_spec into two files
