+++ mitmproxy/proxy/layers/modes.py
@@ -2,14 +2,16 @@ import socket
 import struct
 from abc import ABCMeta
 from dataclasses import dataclass
-from typing import Optional
+from typing import cast, Optional, TYPE_CHECKING
 
 from mitmproxy import connection
 from mitmproxy.proxy import commands, events, layer
 from mitmproxy.proxy.commands import StartHook
 from mitmproxy.proxy.layers import dns, tls
-from mitmproxy.proxy.mode_specs import ReverseMode
 from mitmproxy.proxy.utils import expect
+
+if TYPE_CHECKING:
+    from mitmproxy.proxy.mode_specs import ReverseMode
 
 
 class HttpProxy(layer.Layer):
@@ -55,8 +57,7 @@ class DestinationKnown(layer.Layer, metaclass=ABCMeta):
 class ReverseProxy(DestinationKnown):
     @expect(events.Start)
     def _handle_event(self, event: events.Event) -> layer.CommandGenerator[None]:
-        spec = self.context.client.proxy_mode
-        assert isinstance(spec, ReverseMode)
+        spec = cast(ReverseMode, self.context.client.proxy_mode)
         self.context.server.address = spec.address
 
         if spec.scheme == "https" or spec.scheme == "tls" or spec.scheme == "dtls":
